'''Interpret HIV'''
 * Make classes inheriting from Policy for the reference and discovered policies
 * Compute action regrets using rollouts
 * Make cost-sensitive classification dataset

+++ projects/Proteins/Rosetta/WorkUnit/ProteinOptimizationWorkUnit.h
@@ -274,7 +274,7 @@ public:
 
     workUnits = new CompositeWorkUnit(T("Optimization"));
 
-    for (size_t i = 0; i < inputProteins.size(); i++)
+    for (int i = 0; i < inputProteins.size(); i++)
     {
       ProteinPtr currentProtein = Protein::createFromPDB(context, (*inputProteins[i]));
